+++ weonhyeok/spring-study/src/main/java/spring_study/spring_study/dto/AnswerRequest.java
@@ -1,9 +1,9 @@
 package spring_study.spring_study.dto;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 public record AnswerRequest(
     Long id,
     String content,
-    LocalDate createDate
+    LocalDateTime createDate
 ) {}
