+++ weonhyeok/spring-study/src/main/java/spring_study/spring_study/exception/GlobalQuestionExceptionHandler.java
@@ -1,10 +1,10 @@
 package spring_study.spring_study.exception;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-@Controller
+@ControllerAdvice
 public class GlobalQuestionExceptionHandler {
     @ExceptionHandler(QuestionException.class)
     public String handleMemberException(QuestionException e, Model model) {
